fix comparing of 2 io files

+++ plugins/android/src/org/jetbrains/android/compiler/AndroidCompileUtil.java
@@ -25,6 +25,7 @@ import com.intellij.openapi.compiler.CompileContext;
 import com.intellij.openapi.compiler.CompileScope;
 import com.intellij.openapi.compiler.CompilerMessageCategory;
 import com.intellij.openapi.compiler.GeneratingCompiler;
+import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.module.ModuleManager;
 import com.intellij.openapi.project.Project;
@@ -47,6 +48,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -55,6 +57,8 @@ import java.util.regex.Pattern;
  * @author yole
  */
 public class AndroidCompileUtil {
+  private static final Logger LOG = Logger.getInstance("#org.jetbrains.android.compiler.AndroidCompileUtil");
+
   private static final Pattern ourMessagePattern = Pattern.compile("(.+):(\\d+):.+");
 
   private AndroidCompileUtil() {
@@ -287,7 +291,7 @@ public class AndroidCompileUtil {
   // must be invoked in a read action!
 
   public static void removeDuplicatingClasses(final Module module, @NotNull final String packageName, @NotNull String className,
-                                              @Nullable final File classFile, String sourceRootPath) {
+                                              @Nullable File classFile, String sourceRootPath) {
     if (sourceRootPath == null) {
       return;
     }
@@ -307,17 +311,25 @@ public class AndroidCompileUtil {
         if (virtualFile != null && projectFileIndex.getSourceRootForFile(virtualFile) == sourceRoot) {
           final String path = virtualFile.getPath();
           File f = new File(path);
-          if (!f.equals(classFile) && f.exists()) {
-            if (f.delete()) {
-              virtualFile.refresh(true, false);
+
+          try {
+            f = f.getCanonicalFile();
+            classFile = classFile != null ? classFile.getCanonicalFile() : null;
+            if (f != null && !f.equals(classFile) && f.exists()) {
+              if (f.delete()) {
+                virtualFile.refresh(true, false);
+              }
+              else {
+                ApplicationManager.getApplication().invokeLater(new Runnable() {
+                  public void run() {
+                    Messages.showErrorDialog(project, "Can't delete file " + path, CommonBundle.getErrorTitle());
+                  }
+                }, project.getDisposed());
+              }
             }
-            else {
-              ApplicationManager.getApplication().invokeLater(new Runnable() {
-                public void run() {
-                  Messages.showErrorDialog(project, "Can't delete file " + path, CommonBundle.getErrorTitle());
-                }
-              }, project.getDisposed());
-            }
+          }
+          catch (IOException e) {
+            LOG.info(e);
           }
         }
       }
